requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
gunicorn==20.1.0
pymongo==3.12.0
redis==3.5.3
sentence-transformers==2.2.0
transformers==4.12.5
torch==1.9.0
faiss-cpu==1.7.2
numpy==1.21.2
prometheus-client==0.11.0
python-dotenv==0.19.0
slowapi==0.1.5